Wrap an abstraction around hands.
Previously we were maintaining parallel arrays for hands and we
decided this wasn't going to be nice long term. Wrap an
abstraction around hands so that we are forced to fix all the
call sites, write tests against the abstraction, then replace
the data structure with something that still passes those tests.