Large change to track all possible plays.
Improve inference by tracking all possible values for every
card, assuming good touch and clues as the main filter to get
the list down to manageable sizes. Doesn't keep separate the
"sure information" from the "implicit information" but simply
filters a single list of possible plays down from the universe
of 25 cards to the set that can be played based on the clues
given so far and the conventions from H group.